+++ src/main/java/io/supertokens/storage/mysql/queries/PasswordlessQueries.java
@@ -26,6 +26,7 @@ import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
 import io.supertokens.pluginInterface.multitenancy.TenantIdentifier;
 import io.supertokens.pluginInterface.passwordless.PasswordlessCode;
 import io.supertokens.pluginInterface.passwordless.PasswordlessDevice;
+import io.supertokens.pluginInterface.passwordless.PasswordlessImportUser;
 import io.supertokens.pluginInterface.sqlStorage.SQLStorage.TransactionIsolationLevel;
 import io.supertokens.storage.mysql.ConnectionPool;
 import io.supertokens.storage.mysql.Start;
@@ -35,6 +36,7 @@ import io.supertokens.storage.mysql.utils.Utils;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.sql.Connection;
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.*;
@@ -154,6 +156,75 @@ public class PasswordlessQueries {
         }, TransactionIsolationLevel.REPEATABLE_READ);
     }
 
+    public static void importUsers_Transaction(Connection sqlCon, Start start,
+                                               Collection<PasswordlessImportUser> users) throws SQLException {
+
+        String app_id_to_user_id_QUERY = "INSERT INTO " + getConfig(start).getAppIdToUserIdTable()
+                + "(app_id, user_id, primary_or_recipe_user_id, recipe_id)" + " VALUES(?, ?, ?, ?)";
+        PreparedStatement appIdToUserIdStatement = sqlCon.prepareStatement(app_id_to_user_id_QUERY);
+
+        String all_auth_recipe_users_QUERY = "INSERT INTO " + getConfig(start).getUsersTable() +
+                "(app_id, tenant_id, user_id, primary_or_recipe_user_id, recipe_id, time_joined, " +
+                "primary_or_recipe_user_time_joined)" +
+                " VALUES(?, ?, ?, ?, ?, ?, ?)";
+        PreparedStatement allAuthRecipeUsersStatement = sqlCon.prepareStatement(all_auth_recipe_users_QUERY);
+
+        String passwordless_users_QUERY = "INSERT INTO " + getConfig(start).getPasswordlessUsersTable()
+                + "(app_id, user_id, email, phone_number, time_joined)" + " VALUES(?, ?, ?, ?, ?)";
+        PreparedStatement passwordlessUsersStatement = sqlCon.prepareStatement(passwordless_users_QUERY);
+
+        String passwordless_user_to_tenant_QUERY = "INSERT INTO " + getConfig(start).getPasswordlessUserToTenantTable()
+                + "(app_id, tenant_id, user_id, email, phone_number)" + " VALUES(?, ?, ?, ?, ?)";
+        PreparedStatement passwordlessUserToTenantStatement = sqlCon.prepareStatement(passwordless_user_to_tenant_QUERY);
+
+        int counter = 0;
+        for (PasswordlessImportUser user: users){
+            TenantIdentifier tenantIdentifier = user.tenantIdentifier;
+            appIdToUserIdStatement.setString(1, tenantIdentifier.getAppId());
+            appIdToUserIdStatement.setString(2, user.userId);
+            appIdToUserIdStatement.setString(3, user.userId);
+            appIdToUserIdStatement.setString(4, PASSWORDLESS.toString());
+            appIdToUserIdStatement.addBatch();
+
+            allAuthRecipeUsersStatement.setString(1, tenantIdentifier.getAppId());
+            allAuthRecipeUsersStatement.setString(2, tenantIdentifier.getTenantId());
+            allAuthRecipeUsersStatement.setString(3, user.userId);
+            allAuthRecipeUsersStatement.setString(4, user.userId);
+            allAuthRecipeUsersStatement.setString(5, PASSWORDLESS.toString());
+            allAuthRecipeUsersStatement.setLong(6, user.timeJoinedMSSinceEpoch);
+            allAuthRecipeUsersStatement.setLong(7, user.timeJoinedMSSinceEpoch);
+            allAuthRecipeUsersStatement.addBatch();
+
+            passwordlessUsersStatement.setString(1, tenantIdentifier.getAppId());
+            passwordlessUsersStatement.setString(2, user.userId);
+            passwordlessUsersStatement.setString(3, user.email);
+            passwordlessUsersStatement.setString(4, user.phoneNumber);
+            passwordlessUsersStatement.setLong(5, user.timeJoinedMSSinceEpoch);
+            passwordlessUsersStatement.addBatch();
+
+            passwordlessUserToTenantStatement.setString(1, tenantIdentifier.getAppId());
+            passwordlessUserToTenantStatement.setString(2, tenantIdentifier.getTenantId());
+            passwordlessUserToTenantStatement.setString(3, user.userId);
+            passwordlessUserToTenantStatement.setString(4, user.email);
+            passwordlessUserToTenantStatement.setString(5, user.phoneNumber);
+            passwordlessUserToTenantStatement.addBatch();
+
+            counter++;
+
+            if(counter % 100 == 0) {
+                appIdToUserIdStatement.executeBatch();
+                allAuthRecipeUsersStatement.executeBatch();
+                passwordlessUsersStatement.executeBatch();
+                passwordlessUserToTenantStatement.executeBatch();
+            }
+        }
+
+        appIdToUserIdStatement.executeBatch();
+        allAuthRecipeUsersStatement.executeBatch();
+        passwordlessUsersStatement.executeBatch();
+        passwordlessUserToTenantStatement.executeBatch();
+
+    }
     public static PasswordlessDevice getDevice_Transaction(Start start, Connection con,
                                                            TenantIdentifier tenantIdentifier, String deviceIdHash)
             throws StorageQueryException, SQLException {