+++ src/main/java/io/supertokens/storage/mysql/queries/SessionQueries.java
@@ -26,13 +26,16 @@ import io.supertokens.pluginInterface.multitenancy.TenantIdentifier;
 import io.supertokens.pluginInterface.session.SessionInfo;
 import io.supertokens.storage.mysql.Start;
 import io.supertokens.storage.mysql.config.Config;
+import io.supertokens.storage.mysql.utils.Utils;
 
 import javax.annotation.Nullable;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static io.supertokens.storage.mysql.QueryExecutorTemplate.execute;
 import static io.supertokens.storage.mysql.QueryExecutorTemplate.update;
@@ -390,6 +393,31 @@ public class SessionQueries {
         update(start, QUERY, pst -> {
             pst.setString(1, appIdentifier.getAppId());
             pst.setLong(2, time);
+        });
+    }
+
+    public static Map<String, List<String>> getAllNonExpiredSessionHandlesForUsers(Start start, AppIdentifier appIdentifier,
+                                                                                   List<String> userIds)
+            throws SQLException, StorageQueryException {
+        String QUERY = "SELECT user_id, session_handle FROM " + Config.getConfig(start).getSessionInfoTable()
+                + " WHERE app_id = ? AND expires_at >= ? AND user_id IN ( " + Utils.generateCommaSeperatedQuestionMarks(userIds.size()) + " )";
+
+        return execute(start, QUERY, pst -> {
+            pst.setString(1, appIdentifier.getAppId());
+            pst.setLong(2, currentTimeMillis());
+            for(int i = 0; i < userIds.size() ; i++){
+                pst.setString(3 + i, userIds.get(i));
+            }
+        }, result -> {
+            Map<String, List<String>> temp = new HashMap<>();
+            while (result.next()) {
+                String userId = result.getString("user_id");
+                if(!temp.containsKey(userId)){
+                    temp.put(userId, new ArrayList<>());
+                }
+                temp.get(userId).add(result.getString("session_handle"));
+            }
+            return temp;
         });
     }