+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+- Adds queries for Bulk Import
+- Adds support for multithreaded bulk import
+
 ## [7.2.0] - 2024-10-03
 
 - Compatible with plugin interface version 6.3